requirements.txt
tensorflow>=2.7.0
rasa-open-source
spacy
nltk
transformers
librosa
opencv-python
google-cloud-vision
google-cloud-speech
requests
python-dotenv
scikit-learn
matplotlib
seaborn
streamlit
sounddevice
soundfile
numpy
Pillow
torch
torchvision